+++ routes/api/races.js
@@ -6,6 +6,11 @@ const RaceModel = require('../../models/Race')
 const CheckinModel = require('../../models/Checkin')
 
 router.post('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RaceModel.create(req.body)
 
   res.setHeader('Content-Type', 'application/json')
@@ -13,6 +18,11 @@ router.post('/', async(req, res) => {
 })
 
 router.get('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RaceModel.find()
   res.setHeader('Content-Type', 'application/json')
   res.end(JSON.stringify(result))
@@ -20,6 +30,11 @@ router.get('/', async(req, res) => {
 })
 
 router.get('/:id', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RaceModel.findById(req.params.id)
 
   res.setHeader('Content-Type', 'application/json')
@@ -28,6 +43,11 @@ router.get('/:id', async(req, res) => {
 })
 
 router.put('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RaceModel.findByIdAndUpdate(req.params.id, req.body)
 
   res.setHeader('Content-Type', 'application/json')
@@ -36,6 +56,11 @@ router.put('/', async(req, res) => {
 })
 
 router.delete('/:id', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let tags = await CheckinModel.deleteMany({ race: req.params.id })
   let race = await RaceModel.findByIdAndDelete(req.params.id)
