+++ routes/api/runner.js
@@ -6,6 +6,11 @@ const RunnerModel = require('../../models/Runner')
 const CheckinModel = require('../../models/Checkin')
 
 router.post('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RunnerModel.create(req.body)
 
   res.setHeader('Content-Type', 'application/json')
@@ -13,6 +18,11 @@ router.post('/', async(req, res) => {
 })
 
 router.get('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RunnerModel.find()
 
   res.setHeader('Content-Type', 'application/json')
@@ -21,6 +31,11 @@ router.get('/', async(req, res) => {
 })
 
 router.get('/:id', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RunnerModel.find(req.params.id)
 
   res.setHeader('Content-Type', 'application/json')
@@ -29,6 +44,11 @@ router.get('/:id', async(req, res) => {
 })
 
 router.put('/', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let result = await RunnerModel.findByIdAndUpdate(req.params.id, req.body)
 
   res.setHeader('Content-Type', 'application/json')
@@ -37,6 +57,11 @@ router.put('/', async(req, res) => {
 })
 
 router.delete('/:id', async(req, res) => {
+
+  if (!req.user) {
+    return res.status(500).send({ error: 'Not logged in' })
+  }
+
   let tags = await CheckinModel.deleteMany({ runner: req.params.id })
   let runner = await RunnerModel.findByIdAndDelete(req.params.id)
